@@ -1,36 +1,25 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Guard;
 
 use Dot\Controller\AbstractController;
 use Dot\Rbac\Guard\Exception\RuntimeException;
 use Dot\Rbac\Role\RoleServiceInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Mezzio\Router\RouteResult;
+use Psr\Http\Message\ServerRequestInterface;
 
-/**
- * Class ControllerGuard
- * @package Dot\Rbac\Guard\Controller
- */
+use function in_array;
+use function strtolower;
+
 class ControllerGuard extends AbstractGuard
 {
-    const PRIORITY = 40;
+    public const PRIORITY = 40;
 
-    /** @var RoleServiceInterface */
-    protected $roleService;
+    protected ?RoleServiceInterface $roleService = null;
 
-    /**
-     * ControllerGuard constructor.
-     * @param array $options
-     */
-    public function __construct(array $options = null)
+    public function __construct(?array $options = null)
     {
         $options = $options ?? [];
         parent::__construct($options);
@@ -38,22 +27,19 @@ class ControllerGuard extends AbstractGuard
             $this->setRoleService($options['role_service']);
         }
 
-        if (!$this->roleService instanceof RoleServiceInterface) {
+        if (! $this->roleService instanceof RoleServiceInterface) {
             throw new RuntimeException('RoleService is required by this guard and was not set');
         }
     }
 
-    /**
-     * @param array $rules
-     */
-    public function setRules(array $rules)
+    public function setRules(array $rules): void
     {
         $this->rules = [];
 
         foreach ($rules as $rule) {
-            $route = strtolower($rule['route']);
-            $actions = isset($rule['actions']) ? (array)$rule['actions'] : [];
-            $roles = (array)$rule['roles'];
+            $route   = strtolower($rule['route']);
+            $actions = isset($rule['actions']) ? (array) $rule['actions'] : [];
+            $roles   = (array) $rule['roles'];
 
             if (empty($actions)) {
                 $this->rules[$route][0] = $roles;
@@ -61,47 +47,37 @@ class ControllerGuard extends AbstractGuard
             }
 
             foreach ($actions as $action) {
-                $action = AbstractController::getMethodFromAction($action);
+                $action                       = AbstractController::getMethodFromAction($action);
                 $this->rules[$route][$action] = $roles;
             }
         }
     }
 
-    /**
-     * @return RoleServiceInterface
-     */
-    public function getRoleService(): RoleServiceInterface
+    public function getRoleService(): ?RoleServiceInterface
     {
         return $this->roleService;
     }
 
-    /**
-     * @param RoleServiceInterface $roleService
-     */
-    public function setRoleService(RoleServiceInterface $roleService)
+    public function setRoleService(RoleServiceInterface $roleService): void
     {
         $this->roleService = $roleService;
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @return bool
-     */
     public function isGranted(ServerRequestInterface $request): bool
     {
-        $routeResult = $request->getAttribute(RouteResult::class, null);
-        if (!$routeResult instanceof RouteResult || !$routeResult->getMatchedRouteName()) {
+        $routeResult = $request->getAttribute(RouteResult::class);
+        if (! $routeResult instanceof RouteResult || ! $routeResult->getMatchedRouteName()) {
             return $this->protectionPolicy === self::POLICY_ALLOW;
         }
 
         $route = strtolower($routeResult->getMatchedRouteName());
 
-        if (!isset($this->rules[$route])) {
+        if (! isset($this->rules[$route])) {
             return $this->protectionPolicy === self::POLICY_ALLOW;
         }
 
         $params = $routeResult->getMatchedParams();
-        $action = isset($params['action']) && !empty($params['action'])
+        $action = ! empty($params['action'])
             ? $params['action']
             : 'index';
 
