@@ -1,40 +1,25 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Options;
 
 use Dot\Rbac\Guard\Guard\GuardInterface;
 use Laminas\Stdlib\AbstractOptions;
+use Traversable;
 
-/**
- * Class RbacGuardOptions
- * @package Dot\Rbac\Guard\Options
- */
 class RbacGuardOptions extends AbstractOptions
 {
-    /**
-     * @var string
-     */
-    protected $protectionPolicy = GuardInterface::POLICY_ALLOW;
+    protected string $protectionPolicy = GuardInterface::POLICY_ALLOW;
 
-    /** @var array */
-    protected $eventListeners = [];
+    protected array $eventListeners = [];
 
-    /** @var array */
-    protected $guardsProvider = [];
+    protected array $guardsProvider = [];
 
-    /** @var  MessagesOptions */
-    protected $messagesOptions;
+    protected ?MessagesOptions $messagesOptions = null;
 
     /**
-     * ModuleOptions constructor.
-     * @param array|null|\Traversable $options
+     * @param array|null|Traversable $options
      */
     public function __construct($options)
     {
@@ -42,69 +27,46 @@ class RbacGuardOptions extends AbstractOptions
         parent::__construct($options);
     }
 
-    /**
-     * @return string
-     */
     public function getProtectionPolicy(): string
     {
         return $this->protectionPolicy;
     }
 
-    /**
-     * @param string $protectionPolicy
-     */
-    public function setProtectionPolicy(string $protectionPolicy)
+    public function setProtectionPolicy(string $protectionPolicy): void
     {
         $this->protectionPolicy = $protectionPolicy;
     }
 
-    /**
-     * @return array
-     */
     public function getGuardsProvider(): array
     {
         return $this->guardsProvider;
     }
 
-    /**
-     * @param array $guardsProvider
-     */
-    public function setGuardsProvider(array $guardsProvider)
+    public function setGuardsProvider(array $guardsProvider): void
     {
         $this->guardsProvider = $guardsProvider;
     }
 
-    /**
-     * @return MessagesOptions
-     */
-    public function getMessagesOptions(): MessagesOptions
+    public function getMessagesOptions(): ?MessagesOptions
     {
-        if (!$this->messagesOptions) {
+        if (empty($this->messagesOptions)) {
             $this->setMessagesOptions([]);
         }
+
         return $this->messagesOptions;
     }
 
-    /**
-     * @param array $messagesOptions
-     */
-    public function setMessagesOptions(array $messagesOptions)
+    public function setMessagesOptions(array $messagesOptions): void
     {
         $this->messagesOptions = new MessagesOptions($messagesOptions);
     }
 
-    /**
-     * @return array
-     */
     public function getEventListeners(): array
     {
         return $this->eventListeners;
     }
 
-    /**
-     * @param array $eventListeners
-     */
-    public function setEventListeners(array $eventListeners)
+    public function setEventListeners(array $eventListeners): void
     {
         $this->eventListeners = $eventListeners;
     }
