@@ -1,33 +1,24 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Provider;
 
 use Dot\Rbac\Guard\Guard\GuardInterface;
 
-/**
- * Class GuardsProvider
- * @package Dot\Rbac\Guard\Provider
- */
+use function is_array;
+use function usort;
+
 class ArrayGuardsProvider extends AbstractGuardsProvider
 {
-    /** @var array */
-    protected $guardsConfig = [];
+    protected array $guardsConfig = [];
 
-    /** @var GuardInterface[] */
-    protected $guards;
+    protected array $guards = [];
 
     /**
-     * ArrayGuardsProvider constructor.
-     * @param array $options
+     * @param array|null $options
      */
-    public function __construct(array $options = null)
+    public function __construct(?array $options = null)
     {
         $options = $options ?? [];
         parent::__construct($options);
@@ -38,7 +29,8 @@ class ArrayGuardsProvider extends AbstractGuardsProvider
     }
 
     /**
-     * Gets the  cached guard list or creates it from the config
+     * Gets the cached guard list or creates it from the config
+     *
      * @return GuardInterface[]
      */
     public function getGuards(): array
@@ -62,28 +54,20 @@ class ArrayGuardsProvider extends AbstractGuardsProvider
 
     /**
      * Sort the guards list internally
-     *
-     * @return void
      */
-    protected function sortGuardsByPriority()
+    protected function sortGuardsByPriority(): void
     {
         usort($this->guards, function (GuardInterface $a, GuardInterface $b) {
             return $b->getPriority() - $a->getPriority();
         });
     }
 
-    /**
-     * @return array
-     */
     public function getGuardsConfig(): array
     {
         return $this->guardsConfig;
     }
 
-    /**
-     * @param array $guardsConfig
-     */
-    public function setGuardsConfig(array $guardsConfig)
+    public function setGuardsConfig(array $guardsConfig): void
     {
         $this->guardsConfig = $guardsConfig;
     }
