@@ -1,46 +1,26 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Provider;
 
 use Dot\Rbac\Guard\Exception\RuntimeException;
 use Psr\Container\ContainerInterface;
 
-/**
- * Class Factory
- * @package Dot\Rbac\Guard\Provider
- */
 class Factory
 {
-    /** @var  ContainerInterface */
-    protected $container;
+    protected ContainerInterface $container;
 
-    /** @var  GuardsProviderPluginManager */
-    protected $guardsProviderPluginManager;
+    protected ?GuardsProviderPluginManager $guardsProviderPluginManager;
 
-    /**
-     * Factory constructor.
-     * @param ContainerInterface $container
-     * @param GuardsProviderPluginManager|null $guardsProviderPluginManager
-     */
     public function __construct(
         ContainerInterface $container,
-        GuardsProviderPluginManager $guardsProviderPluginManager = null
+        ?GuardsProviderPluginManager $guardsProviderPluginManager = null
     ) {
-        $this->container = $container;
+        $this->container                   = $container;
         $this->guardsProviderPluginManager = $guardsProviderPluginManager;
     }
 
-    /**
-     * @param array $specs
-     * @return GuardsProviderInterface
-     */
     public function create(array $specs): GuardsProviderInterface
     {
         $type = $specs['type'] ?? '';
@@ -51,12 +31,9 @@ class Factory
         return $this->getGuardsProviderPluginManager()->get($type, $specs['options'] ?? null);
     }
 
-    /**
-     * @return GuardsProviderPluginManager
-     */
     public function getGuardsProviderPluginManager(): GuardsProviderPluginManager
     {
-        if (!$this->guardsProviderPluginManager) {
+        if (! $this->guardsProviderPluginManager) {
             $this->guardsProviderPluginManager = new GuardsProviderPluginManager($this->container, []);
         }
 
