@@ -1,24 +1,21 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Event;
 
+use ArrayAccess;
 use Dot\Event\Event;
 
 /**
- * Class AuthorizationEvent
- * @package Dot\Rbac\Guard\Event
+ * @template TTarget of object|string|null
+ * @template TParams of array|ArrayAccess|object
+ * @extends Event<TTarget, TParams>
  */
 class AuthorizationEvent extends Event
 {
-    const EVENT_BEFORE_AUTHORIZATION = 'event.beforeAuthorization';
-    const EVENT_AFTER_AUTHORIZATION = 'event.afterAuthorization';
+    public const EVENT_BEFORE_AUTHORIZATION = 'event.beforeAuthorization';
+    public const EVENT_AFTER_AUTHORIZATION  = 'event.afterAuthorization';
 
-    const EVENT_FORBIDDEN = 'event.authorization.forbidden';
+    public const EVENT_FORBIDDEN = 'event.authorization.forbidden';
 }
