@@ -1,26 +1,17 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Event;
 
 use Laminas\EventManager\EventManagerInterface;
 use Laminas\EventManager\ListenerAggregateTrait;
 
-/**
- * Class AuthorizationEventListenerTrait
- * @package Dot\Rbac\Guard\Event
- */
 trait AuthorizationEventListenerTrait
 {
     use ListenerAggregateTrait;
 
-    public function attach(EventManagerInterface $events, $priority = 1)
+    public function attach(EventManagerInterface $events, mixed $priority = 1): void
     {
         $this->listeners[] = $events->attach(
             AuthorizationEvent::EVENT_BEFORE_AUTHORIZATION,
@@ -39,27 +30,15 @@ trait AuthorizationEventListenerTrait
         );
     }
 
-    /**
-     * @param AuthorizationEvent $e
-     */
-    public function onBeforeAuthorization(AuthorizationEvent $e)
+    public function onBeforeAuthorization(AuthorizationEvent $e): void
     {
-        // no-op
     }
 
-    /**
-     * @param AuthorizationEvent $e
-     */
-    public function onAfterAuthorization(AuthorizationEvent $e)
+    public function onAfterAuthorization(AuthorizationEvent $e): void
     {
-        // no-op
     }
 
-    /**
-     * @param AuthorizationEvent $e
-     */
-    public function onForbidden(AuthorizationEvent $e)
+    public function onForbidden(AuthorizationEvent $e): void
     {
-        // no-op
     }
 }
