@@ -1,19 +1,9 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Exception;
 
-/**
- * Class InvalidArgumentException
- * @package Dot\Rbac\Guard\Exception
- */
 class InvalidArgumentException extends \InvalidArgumentException implements ExceptionInterface
 {
-
 }
