@@ -1,11 +1,6 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Factory;
 
@@ -15,29 +10,25 @@ use Dot\Rbac\Guard\Middleware\RbacGuardMiddleware;
 use Dot\Rbac\Guard\Options\RbacGuardOptions;
 use Dot\Rbac\Guard\Provider\Factory;
 use Dot\Rbac\Guard\Provider\GuardsProviderPluginManager;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 
-/**
- * Class RbacGuardMiddlewareFactory
- * @package Dot\Rbac\Guard\Factory
- */
 class RbacGuardMiddlewareFactory
 {
     use AttachAuthorizationEventListenersTrait;
 
     /**
-     * @param ContainerInterface $container
-     * @param $requestedName
-     * @return RbacGuardMiddleware
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
      */
-    public function __invoke(ContainerInterface $container, $requestedName)
+    public function __invoke(ContainerInterface $container, string $requestedName): RbacGuardMiddleware
     {
         /** @var RbacGuardOptions $options */
         $options = $container->get(RbacGuardOptions::class);
 
-        /** @var Factory $guardsProviderFactory */
         $guardsProviderFactory = new Factory($container, $container->get(GuardsProviderPluginManager::class));
-        $guardsProvider = $guardsProviderFactory->create($options->getGuardsProvider());
+        $guardsProvider        = $guardsProviderFactory->create($options->getGuardsProvider());
 
         /** @var RbacGuardMiddleware $middleware */
         $middleware = new $requestedName(
