@@ -1,29 +1,20 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-rbac-guard/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-rbac-guard/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Rbac\Guard\Factory;
 
-use Dot\Rbac\Guard\Options\RbacGuardOptions;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 
-/**
- * Class RbacGuardOptionsFactory
- * @package Dot\Rbac\Guard\Factory
- */
 class RbacGuardOptionsFactory
 {
     /**
-     * @param ContainerInterface $container
-     * @param $requestedName
-     * @return RbacGuardOptions
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
      */
-    public function __invoke(ContainerInterface $container, $requestedName)
+    public function __invoke(ContainerInterface $container, string $requestedName): mixed
     {
         return new $requestedName($container->get('config')['dot_authorization']);
     }
